pyproject.toml
[project]
name = "pretraining-mllm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.0.0",
    "pillow>=11.3.0",
    "torch>=2.7.1",
    "transformers>=4.54.0",
]
